requirements.txt

python-pptx
